# schemas/openapi.yaml
openapi: 3.0.3
info:
  title: Books API
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
  /books:
    get:
      operationId: getBooks
      summary: List all books
      tags:
        - Books
      responses:
        '200':
          description: Array of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
    post:
      operationId: createBook
      summary: Create a book
      tags:
        - Books
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBook'
      responses:
        '201':
          description: The created book
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
  /books/{id}:
    get:
      operationId: getBookById
      summary: Get a single book by id
      tags:
        - Books
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The book
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: deleteBook
      summary: Delete a book by id
      tags:
        - Books
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The deleted book
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      operationId: updateBook
      summary: Replace a book by id
      tags:
        - Books
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '200':
          description: The updated book
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Book:
      type: object
      required:
        - title
        - author
        - pages
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: The Shining
        author:
          type: string
          example: Stephen King
        pages:
          type: integer
          example: 447
    CreateBook:
      type: object
      required:
        - title
        - author
        - pages
      properties:
        title:
          type: string
          example: The Shining
        author:
          type: string
          example: Stephen King
        pages:
          type: integer
          example: 447
    Error:
      type: object
      properties:
        message:
          type: string
tags: []